requirements.txt
alembic==1.7.7
Flask==1.1.4
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
flask-restx==0.4.0
Flask-SQLAlchemy==2.5.1
flask-marshmallow[sqlalchemy]==0.14.0
Jinja2==2.11.3
MarkupSafe==0.23
marshmallow-sqlalchemy=0.29.0
python-dotenv==0.18.0
SQLAlchemy==1.4.22